udemy_langgraph/pyproject.toml
[project]
name = "udemy-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10, <3.11"
dependencies = [
    "langchain-community>=0.3.17",
    "langchain-core>=0.3.34",
    "langchain-openai>=0.3.4",
    "langgraph>=0.2.70",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]

[project.scripts]
reflection = "reflection_agent.main:main"
reflexion = "reflexion_agent.main:main"

[tool.uv]
package = true

[tool.uv.workspace]
members = ["reflection_agent","reflexion_agent"]

[tool.setuptools]
packages = ["reflection_agent", "reflexion_agent"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"